backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-jwt-extended==4.6.0
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
python-dotenv==1.0.1
marshmallow==3.21.3
pytest==8.2.2
pytest-flask==1.3.0
anthropic==0.28.0
python-dateutil==2.9.0
